Moved package/errata sub-commands from content-host to host
