+++ tests/foreman/cli/test_contenthost.py
@@ -5,8 +5,6 @@ from robottelo.cli.base import CLIReturnCodeError
 from robottelo.cli.factory import (
     CLIFactoryError,
     make_activation_key,
-    setup_org_for_a_custom_repo,
-    setup_org_for_a_rh_repo,
     make_content_host,
     make_content_view,
     make_lifecycle_environment,
@@ -15,25 +13,10 @@ from robottelo.cli.factory import (
 from robottelo.cli.contenthost import ContentHost
 from robottelo.cli.contentview import ContentView
 from robottelo.cli.lifecycleenvironment import LifecycleEnvironment
-from robottelo.constants import (
-    FAKE_0_CUSTOM_PACKAGE,
-    FAKE_0_CUSTOM_PACKAGE_GROUP,
-    FAKE_0_CUSTOM_PACKAGE_GROUP_NAME,
-    FAKE_0_CUSTOM_PACKAGE_NAME,
-    FAKE_1_CUSTOM_PACKAGE,
-    FAKE_1_CUSTOM_PACKAGE_NAME,
-    FAKE_2_CUSTOM_PACKAGE,
-    FAKE_0_ERRATA_ID,
-    FAKE_0_YUM_REPO,
-    PRDS,
-    REPOS,
-    REPOSET,
-)
 from robottelo.datafactory import invalid_values_list, generate_strings_list
 from robottelo.decorators import (
     run_only_on,
     skip_if_bug_open,
-    skip_if_not_set,
     tier1,
     tier3,
 )
@@ -502,251 +485,3 @@ class ContentHostTestCase(CLITestCase):
             result = ContentHost.list({'organization-id': self.NEW_ORG['id']})
             self.assertNotIn(
                 client.hostname, [chost['name'] for chost in result])
-
-
-class KatelloAgentTestCase(CLITestCase):
-    """Content-host tests, which require VM with installed katello-agent."""
-
-    org = None
-    env = None
-    content_view = None
-    activation_key = None
-
-    @classmethod
-    @skip_if_not_set('clients', 'fake_manifest')
-    def setUpClass(cls):
-        """Create Org, Lifecycle Environment, Content View, Activation key
-
-        """
-        super(KatelloAgentTestCase, cls).setUpClass()
-        # Create new org, environment, CV and activation key
-        KatelloAgentTestCase.org = make_org()
-        KatelloAgentTestCase.env = make_lifecycle_environment({
-            u'organization-id': KatelloAgentTestCase.org['id'],
-        })
-        KatelloAgentTestCase.content_view = make_content_view({
-            u'organization-id': KatelloAgentTestCase.org['id'],
-        })
-        KatelloAgentTestCase.activation_key = make_activation_key({
-            u'lifecycle-environment-id': KatelloAgentTestCase.env['id'],
-            u'organization-id': KatelloAgentTestCase.org['id'],
-        })
-        # Add subscription to Satellite Tools repo to activation key
-        setup_org_for_a_rh_repo({
-            u'product': PRDS['rhel'],
-            u'repository-set': REPOSET['rhst7'],
-            u'repository': REPOS['rhst7']['name'],
-            u'organization-id': KatelloAgentTestCase.org['id'],
-            u'content-view-id': KatelloAgentTestCase.content_view['id'],
-            u'lifecycle-environment-id': KatelloAgentTestCase.env['id'],
-            u'activationkey-id': KatelloAgentTestCase.activation_key['id'],
-        })
-        # Create custom repo, add subscription to activation key
-        setup_org_for_a_custom_repo({
-            u'url': FAKE_0_YUM_REPO,
-            u'organization-id': KatelloAgentTestCase.org['id'],
-            u'content-view-id': KatelloAgentTestCase.content_view['id'],
-            u'lifecycle-environment-id': KatelloAgentTestCase.env['id'],
-            u'activationkey-id': KatelloAgentTestCase.activation_key['id'],
-        })
-
-    def setUp(self):
-        """Create VM, subscribe it to satellite-tools repo, install katello-ca
-        and katello-agent packages
-
-        """
-        super(KatelloAgentTestCase, self).setUp()
-        # Create VM and register content host
-        self.client = VirtualMachine(distro='rhel71')
-        self.client.create()
-        self.client.install_katello_ca()
-        # Register content host, install katello-agent
-        self.client.register_contenthost(
-            KatelloAgentTestCase.activation_key['name'],
-            KatelloAgentTestCase.org['label']
-        )
-        self.client.enable_repo(REPOS['rhst7']['id'])
-        self.client.install_katello_agent()
-
-    def tearDown(self):
-        self.client.destroy()
-        super(KatelloAgentTestCase, self).tearDown()
-
-    @tier3
-    @run_only_on('sat')
-    def test_positive_get_errata_info(self):
-        """Get errata info
-
-        @Feature: Content Host - Errata
-
-        @Assert: Errata info was displayed
-
-        """
-        self.client.download_install_rpm(
-            FAKE_0_YUM_REPO,
-            FAKE_0_CUSTOM_PACKAGE
-        )
-        result = ContentHost.errata_info({
-            u'content-host': self.client.hostname,
-            u'id': FAKE_0_ERRATA_ID,
-            u'organization-id': KatelloAgentTestCase.org['id'],
-        })
-        self.assertEqual(result[0]['errata-id'], FAKE_0_ERRATA_ID)
-        self.assertEqual(result[0]['packages'], FAKE_0_CUSTOM_PACKAGE)
-
-    @tier3
-    @run_only_on('sat')
-    def test_positive_apply_errata(self):
-        """Apply errata to content host
-
-        @Feature: Content Host - Errata
-
-        @Assert: Errata is scheduled for installation
-
-        """
-        self.client.download_install_rpm(
-            FAKE_0_YUM_REPO,
-            FAKE_0_CUSTOM_PACKAGE
-        )
-        ContentHost.errata_apply({
-            u'content-host': self.client.hostname,
-            u'errata-ids': FAKE_0_ERRATA_ID,
-            u'organization-id': KatelloAgentTestCase.org['id'],
-        })
-
-    @tier3
-    @run_only_on('sat')
-    def test_positive_install_package(self):
-        """Install package to content host remotely
-
-        @Feature: Content Host - Package
-
-        @Assert: Package was successfully installed
-
-        """
-        ContentHost.package_install({
-            u'content-host': self.client.hostname,
-            u'organization-id': KatelloAgentTestCase.org['id'],
-            u'packages': FAKE_0_CUSTOM_PACKAGE_NAME,
-        })
-        result = self.client.run(
-            'rpm -q {0}'.format(FAKE_0_CUSTOM_PACKAGE_NAME)
-        )
-        self.assertEqual(result.return_code, 0)
-
-    @tier3
-    @run_only_on('sat')
-    def test_positive_remove_package(self):
-        """Remove package from content host remotely
-
-        @Feature: Content Host - Package
-
-        @Assert: Package was successfully removed
-
-        """
-        self.client.download_install_rpm(
-            FAKE_0_YUM_REPO,
-            FAKE_0_CUSTOM_PACKAGE
-        )
-        ContentHost.package_remove({
-            u'content-host': self.client.hostname,
-            u'organization-id': KatelloAgentTestCase.org['id'],
-            u'packages': FAKE_0_CUSTOM_PACKAGE_NAME,
-        })
-        result = self.client.run(
-            'rpm -q {0}'.format(FAKE_0_CUSTOM_PACKAGE_NAME)
-        )
-        self.assertNotEqual(result.return_code, 0)
-
-    @tier3
-    @run_only_on('sat')
-    def test_positive_upgrade_package(self):
-        """Upgrade content host package remotely
-
-        @Feature: Content Host - Package
-
-        @Assert: Package was successfully upgraded
-
-        """
-        self.client.run('yum install -y {0}'.format(FAKE_1_CUSTOM_PACKAGE))
-        ContentHost.package_upgrade({
-            u'content-host': self.client.hostname,
-            u'organization-id': KatelloAgentTestCase.org['id'],
-            u'packages': FAKE_1_CUSTOM_PACKAGE_NAME,
-        })
-        result = self.client.run('rpm -q {0}'.format(FAKE_2_CUSTOM_PACKAGE))
-        self.assertEqual(result.return_code, 0)
-
-    @tier3
-    @run_only_on('sat')
-    def test_positive_upgrade_packages_all(self):
-        """Upgrade all the content host packages remotely
-
-        @Feature: Content Host - Package
-
-        @Assert: Packages (at least 1 with newer version available) were
-        successfully upgraded
-
-        """
-        self.client.run('yum install -y {0}'.format(FAKE_1_CUSTOM_PACKAGE))
-        ContentHost.package_upgrade_all({
-            u'content-host': self.client.hostname,
-            u'organization-id': KatelloAgentTestCase.org['id'],
-        })
-        result = self.client.run('rpm -q {0}'.format(FAKE_2_CUSTOM_PACKAGE))
-        self.assertEqual(result.return_code, 0)
-
-    @tier3
-    @run_only_on('sat')
-    def test_positive_install_package_group(self):
-        """Install package group to content host remotely
-
-        @Feature: Content Host - Package group
-
-        @Assert: Package group was successfully installed
-
-        """
-        ContentHost.package_group_install({
-            u'content-host': self.client.hostname,
-            u'groups': FAKE_0_CUSTOM_PACKAGE_GROUP_NAME,
-            u'organization-id': KatelloAgentTestCase.org['id'],
-        })
-        for package in FAKE_0_CUSTOM_PACKAGE_GROUP:
-            result = self.client.run('rpm -q {0}'.format(package))
-            self.assertEqual(result.return_code, 0)
-
-    @tier3
-    @run_only_on('sat')
-    def test_positive_remove_package_group(self):
-        """Remove package group from content host remotely
-
-        @Feature: Content Host - Package group
-
-        @Assert: Package group was successfully removed
-
-        """
-        hammer_args = {
-            u'content-host': self.client.hostname,
-            u'groups': FAKE_0_CUSTOM_PACKAGE_GROUP_NAME,
-            u'organization-id': KatelloAgentTestCase.org['id'],
-        }
-        ContentHost.package_group_install(hammer_args)
-        ContentHost.package_group_remove(hammer_args)
-        for package in FAKE_0_CUSTOM_PACKAGE_GROUP:
-            result = self.client.run('rpm -q {0}'.format(package))
-            self.assertNotEqual(result.return_code, 0)
-
-    @tier3
-    def test_negative_unregister_and_pull_content(self):
-        """Attempt to retrieve content after Content host has been
-        unregistered from Satellite
-
-        @feature: Content host
-
-        @assert: Content host can no longer retrieve content from satellite
-        """
-        result = self.client.run('subscription-manager unregister')
-        self.assertEqual(result.return_code, 0)
-        result = self.client.run(
-            'yum install -y {0}'.format(FAKE_1_CUSTOM_PACKAGE))
-        self.assertNotEqual(result.return_code, 0)
